requirements.txt
absl-py==1.2.0
attrs==22.1.0
contourpy==1.0.5
customtkinter==4.6.3
cycler==0.11.0
darkdetect==0.7.1
fonttools==4.37.3
future==0.18.2
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.0
mediapipe==0.8.9.1
numpy==1.23.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
pickle4==0.0.1
Pillow==9.2.0
protobuf==3.20.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tkintertable==1.3.3
